requirements.txt
# Core dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Data processing
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Geospatial libraries for UAV data
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment tracking (optional)
tensorboard>=2.13.0
# wandb>=0.15.0  # Uncomment if using Weights & Biases

# Utils
scipy>=1.10.0
scikit-learn>=1.3.0
